+++ app/helpers/komponent/styleguide_helper.rb
@@ -1,7 +1,0 @@
-module Komponent
-  module StyleguideHelper
-    def komponent_components
-      Komponent::Component.all
-    end
-  end
-end