+++ lib/komponent.rb
@@ -3,6 +3,5 @@
 require 'komponent/version'
 
 module Komponent
-  require 'komponent/railtie' if defined?(Rails)
   require 'komponent/engine' if defined?(Rails)
 end